Moved and activated testcase
